+++ R/RWeka_CBA.R
@@ -2,7 +2,8 @@
 #'
 #' Provides CBA-type classifiers based on RIPPER (Cohen, 1995), C4.5 (Quinlan,
 #' 1993) and PART (Frank and Witten, 1998) using the implementation in Weka via
-#' RWeka (Hornik et al, 2009).
+#' RWeka (Hornik et al, 2009). These classifiers do not mine CARs, but
+#' directly create rules.
 #'
 #' You need to install package \pkg{RWeka} to use these classifiers.
 #'
@@ -15,9 +16,11 @@
 #'
 #' Memory for \pkg{RWeka} can be increased using the R options (e.g.,
 #' `options(java.parameters = "-Xmx1024m")`) before \pkg{RWeka} or
-#' \pkg{rJava} is loaded or any RWeka-based classigier in this package is used.
+#' \pkg{rJava} is loaded or any RWeka-based classifier in this package is used.
 #'
 #' @name RWeka_CBA
+#'
+#' @family classifiers
 #'
 #' @param formula A symbolic description of the model to be fitted. Has to be
 #'   of form `class ~ .` or `class ~ predictor1 + predictor2`.