+++ R/predict.R
@@ -5,6 +5,8 @@
 #' @aliases predict
 #' @name predict.CBA
 #'
+#' @family classifier
+#'
 #' @param object An object of class [CBA].
 #' @param newdata A data.frame or [arules::transactions] containing rows of new entries
 #' to be classified.