+++ R/transactions2DF.R
@@ -3,7 +3,9 @@
 #' Convert transactions back into data.frames by combining the
 #' items for the same variable into a single column.
 #'
-#' @param transactions an object of class [transactions].
+#' @family preparation
+#'
+#' @param transactions an object of class [arules::transactions].
 #' @param itemLabels logical; use the complete item labels (variable=level) as the
 #'    levels in the data.frame? By default, only the levels are used.
 #' @return Returns a data.frame.